modified

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,17 +9,17 @@ int main(void)
 {
 	int counter;
 	int countto = 50;
-	long a = 1;
-	long b = 2;
+	int a = 1;
+	int b = 2;
 
 	for (counter = 1; counter <= (countto / 2); counter++)
 	{
-		printf("%li %li ", a, b);
+		printf("%d %d ", a, b);
 		a += b;
 		b += a;
 	}
 	if (countto % 2 == 1)
-		printf("%li", a);
+		printf("%d", a);
 
 	printf("\n");
